Extract pure sync core into er_engine::sync (TabState-free)
Move the App-independent halves of github_sync.rs and remote_diff_sync.rs
into a new always-on sync module so headless consumers (er-api) can reuse
comment merge/anchor resolution and remote diff fetching:

- fetch_comment_sync_data + CommentSyncContext/Result
- resolve_anchor, find_local_line_for_diff_hunk, extract_anchor_from_diff_hunk
- fetch_remote_diff_data + RemoteDiffContext/Result
- chrono_now (was app::state, now re-exported from there)

app/state/{github_sync,remote_diff_sync}.rs keep the App wrappers
(snapshot/apply) and re-export the moved types, so all existing
crate::app::* paths in er-tui and er-desktop are unchanged.
