+++ crates/er-engine/src/app/state/remote_diff_sync.rs
@@ -10,41 +10,11 @@
 //! `expected_head_oid`; if it matches the tab's `last_diff_head_oid` the
 //! fetch is skipped and `Ok(None)` is returned.
 
-use anyhow::Result;
-
-use crate::ai::{compute_diff_hash, compute_diff_hash_fast};
-use crate::git;
-use crate::github;
+// The pure fetch core (no App dependency) lives in `crate::sync`; re-exported
+// here so existing `crate::app::...` paths keep working.
+pub use crate::sync::{fetch_remote_diff_data, RemoteDiffContext, RemoteDiffResult};
 
 use super::App;
-
-/// Inputs for one refresh cycle. Built while holding the App lock.
-#[derive(Debug, Clone)]
-pub struct RemoteDiffContext {
-    pub owner: String,
-    pub repo: String,
-    pub pr_number: u64,
-    /// Used by `apply_remote_diff_result` to find the right tab if the user
-    /// switches or closes tabs during the network fetch.
-    pub repo_root: String,
-    /// What `last_diff_head_oid` was on the tab when the snapshot was taken.
-    pub last_head_oid: Option<String>,
-    /// Latest head_oid available out-of-band (typically the PR cache). When
-    /// this equals `last_head_oid` the loop short-circuits.
-    pub expected_head_oid: Option<String>,
-}
-
-/// Output of one refresh cycle. Applied via `apply_remote_diff_result`.
-#[derive(Debug, Clone)]
-pub struct RemoteDiffResult {
-    pub raw_diff: String,
-    pub files: Vec<git::DiffFile>,
-    pub branch_diff_hash: String,
-    pub diff_hash: String,
-    pub head_oid: Option<String>,
-    /// (repo_root, pr_number, is_remote) — used to find the right tab on apply.
-    pub tab_key: (String, Option<u64>, bool),
-}
 
 impl App {
     /// Snapshot the active tab's identity for a remote-diff refresh. Returns
@@ -96,32 +66,4 @@ impl App {
             }
         }
     }
-}
-
-/// Run the network fetch + parse + hash for a remote-PR diff refresh.
-/// Returns `Ok(None)` when the expected head_oid matches the last fetched
-/// one (no work needed).
-pub fn fetch_remote_diff_data(ctx: &RemoteDiffContext) -> Result<Option<RemoteDiffResult>> {
-    if let (Some(expected), Some(last)) = (
-        ctx.expected_head_oid.as_deref(),
-        ctx.last_head_oid.as_deref(),
-    ) {
-        if expected == last {
-            return Ok(None);
-        }
-    }
-
-    let raw = github::gh_pr_diff_remote(&ctx.owner, &ctx.repo, ctx.pr_number)?;
-    let files = git::parse_diff(&raw);
-    let branch_diff_hash = compute_diff_hash(&raw);
-    let diff_hash = format!("{:016x}", compute_diff_hash_fast(&raw));
-
-    Ok(Some(RemoteDiffResult {
-        raw_diff: raw,
-        files,
-        branch_diff_hash,
-        diff_hash,
-        head_oid: ctx.expected_head_oid.clone(),
-        tab_key: (ctx.repo_root.clone(), Some(ctx.pr_number), true),
-    }))
 }
